Extract playback progress polling
